requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
fpdf==1.7.2
idna==3.3
img2pdf==0.4.4
lxml==4.9.0
numpy==1.23.0
opencv-python==4.6.0.66
packaging==21.3
pdf2image==1.16.0
pikepdf==5.1.5.post1
Pillow==9.1.1
pyparsing==3.0.9
pytesseract==0.3.9
requests==2.28.0
tqdm==4.64.0
urllib3==1.26.9
termcolor==1.1.0
easyocr==1.5.0